Windows: don't allow fractional DPI scaling

+++ code/Modules/Gfx/win/winDisplayMgr.h
@@ -101,8 +101,8 @@ public:
     DWORD dwExStyle;
     bool inCreateWindow;
     bool dpiAware;
-    float windowScale;
-    float contentScale;
+    int windowScale;
+    int contentScale;
 
     int cursorMode;
     double cursorPosX;
